Provided extra examples of the tee command
The new examples cover two other features of the command
The first is how the -a flag is used. An example is provided and the
output of the command is clearly displayed
The second is how to output to multiple files. Again, an example
is provided, and then the output of the command is clearly displayed